+++ window_scene/interfaces/kits/napi/scene_session_manager/js_scene_session.cpp
@@ -58,8 +58,8 @@ NativeValue* JsSceneSession::Create(NativeEngine& engine, const sptr<SceneSessio
 
     std::unique_ptr<JsSceneSession> jsSceneSession = std::make_unique<JsSceneSession>(engine, session);
     object->SetNativePointer(jsSceneSession.release(), JsSceneSession::Finalizer, nullptr);
-    object->SetProperty("persistentId", CreateJsValue(engine, static_cast<int64_t>(session->GetPersistentId())));
-    object->SetProperty("parentId", CreateJsValue(engine, static_cast<int64_t>(session->GetParentPersistentId())));
+    object->SetProperty("persistentId", CreateJsValue(engine, static_cast<uint32_t>(session->GetPersistentId())));
+    object->SetProperty("parentId", CreateJsValue(engine, static_cast<uint32_t>(session->GetParentPersistentId())));
     object->SetProperty("type", CreateJsValue(engine, static_cast<uint32_t>(GetApiType(session->GetWindowType()))));
     const char* moduleName = "JsSceneSession";
     BindNativeFunction(engine, *object, "on", moduleName, JsSceneSession::RegisterCallback);
