+++ window_scene/session_manager/include/extension_session_manager.h
@@ -63,7 +63,7 @@ private:
     sptr<AAFwk::SessionInfo> SetAbilitySessionInfo(const sptr<ExtensionSession>& extensionSession);
 
     std::shared_ptr<TaskScheduler> taskScheduler_;
-    std::map<uint64_t, sptr<ExtensionSession>> extensionSessionMap_;
+    std::map<uint32_t, sptr<ExtensionSession>> extensionSessionMap_;
 };
 } // namespace OHOS::Rosen
